Main Menu Coverage Testing Files
SO there's many of them because I had to test them manually. Adding all the sections together makes for roughly 95% coverage. It will never be 100% because 2 lines should not ever actually be executed. If they do something has gone very wrong. But all of them have been properly tested in a TDD style.